+++ plugins/scaffolder-backend-module-bitbucket/README.md
@@ -1,0 +1,5 @@
+# @backstage/plugin-scaffolder-backend-module-bitbucket
+
+The bitbucket module for [@backstage/plugin-scaffolder-backend](https://www.npmjs.com/package/@backstage/plugin-scaffolder-backend).
+
+_This plugin was created through the Backstage CLI_
