+++ plugins/scaffolder-backend-module-bitbucket/src/actions/bitbucket.examples.ts
