+++ plugins/scaffolder-backend-module-bitbucket/src/actions/bitbucketServer.test.ts
@@ -14,8 +14,9 @@
  * limitations under the License.
  */
 
-jest.mock('../helpers', () => {
+jest.mock('@backstage/plugin-scaffolder-node', () => {
   return {
+    ...jest.requireActual('@backstage/plugin-scaffolder-node'),
     initRepoAndPush: jest.fn().mockResolvedValue({
       commitHash: '220f19cc36b551763d157f1b5e4a4b446165dbd6',
     }),
@@ -25,7 +26,7 @@ jest.mock('../helpers', () => {
   };
 });
 
-import { createPublishBitbucketAction } from './bitbucket';
+import { createPublishBitbucketServerAction } from './bitbucketServer';
 import { rest } from 'msw';
 import { setupServer } from 'msw/node';
 import { setupRequestMockHandlers } from '@backstage/backend-test-utils';
@@ -33,36 +34,38 @@ import { ScmIntegrations } from '@backstage/integration';
 import { ConfigReader } from '@backstage/config';
 import { getVoidLogger } from '@backstage/backend-common';
 import { PassThrough } from 'stream';
-import { initRepoAndPush } from '../helpers';
+import { initRepoAndPush } from '@backstage/plugin-scaffolder-node';
 
-describe('publish:bitbucket', () => {
+describe('publish:bitbucketServer', () => {
   const config = new ConfigReader({
     integrations: {
-      bitbucket: [
-        {
-          host: 'bitbucket.org',
-          token: 'tokenlols',
-        },
+      bitbucketServer: [
         {
           host: 'hosted.bitbucket.com',
           token: 'thing',
           apiBaseUrl: 'https://hosted.bitbucket.com/rest/api/1.0',
         },
         {
-          host: 'notoken.bitbucket.com',
+          host: 'basic-auth.bitbucket.com',
+          username: 'test-user',
+          password: 'test-password',
+          apiBaseUrl: 'https://basic-auth.bitbucket.com/rest/api/1.0',
+        },
+        {
+          host: 'no-credentials.bitbucket.com',
         },
       ],
     },
   });
 
   const integrations = ScmIntegrations.fromConfig(config);
-  const action = createPublishBitbucketAction({ integrations, config });
+  const action = createPublishBitbucketServerAction({ integrations, config });
   const mockContext = {
     input: {
-      repoUrl: 'bitbucket.org?workspace=workspace&project=project&repo=repo',
+      repoUrl: 'hosted.bitbucket.com?project=project&repo=repo',
       repoVisibility: 'private' as const,
     },
-    workspacePath: 'lol',
+    workspacePath: 'wsp',
     logger: getVoidLogger(),
     logStream: new PassThrough(),
     output: jest.fn(),
@@ -79,21 +82,20 @@ describe('publish:bitbucket', () => {
     await expect(
       action.handler({
         ...mockContext,
-        input: { repoUrl: 'bitbucket.org?project=project&repo=repo' },
-      }),
-    ).rejects.toThrow(/missing workspace/);
-
-    await expect(
-      action.handler({
-        ...mockContext,
-        input: { repoUrl: 'bitbucket.org?workspace=workspace&repo=repo' },
+        input: {
+          ...mockContext.input,
+          repoUrl: 'hosted.bitbucket.com?repo=repo',
+        },
       }),
     ).rejects.toThrow(/missing project/);
 
     await expect(
       action.handler({
         ...mockContext,
-        input: { repoUrl: 'bitbucket.org?workspace=workspace&project=project' },
+        input: {
+          ...mockContext.input,
+          repoUrl: 'hosted.bitbucket.com?project=project',
+        },
       }),
     ).rejects.toThrow(/missing repo/);
   });
@@ -103,74 +105,39 @@ describe('publish:bitbucket', () => {
       action.handler({
         ...mockContext,
         input: {
-          repoUrl: 'missing.com?workspace=workspace&project=project&repo=repo',
+          ...mockContext.input,
+          repoUrl: 'missing.com?project=project&repo=repo',
         },
       }),
     ).rejects.toThrow(/No matching integration configuration/);
   });
 
-  it('should throw if there is no token in the integration config that is returned', async () => {
+  it('should throw if there no credentials in the integration config that is returned', async () => {
     await expect(
       action.handler({
         ...mockContext,
         input: {
-          repoUrl:
-            'notoken.bitbucket.com?workspace=workspace&project=project&repo=repo',
+          ...mockContext.input,
+          repoUrl: 'no-credentials.bitbucket.com?project=project&repo=repo',
         },
       }),
-    ).rejects.toThrow(/Authorization has not been provided for Bitbucket/);
-  });
-
-  it('should call the correct APIs when the host is bitbucket cloud', async () => {
-    expect.assertions(2);
-    server.use(
-      rest.post(
-        'https://api.bitbucket.org/2.0/repositories/workspace/repo',
-        (req, res, ctx) => {
-          expect(req.headers.get('Authorization')).toBe('Bearer tokenlols');
-          expect(req.body).toEqual({
-            is_private: true,
-            scm: 'git',
-            project: { key: 'project' },
-          });
-          return res(
-            ctx.status(200),
-            ctx.set('Content-Type', 'application/json'),
-            ctx.json({
-              links: {
-                html: {
-                  href: 'https://bitbucket.org/workspace/repo',
-                },
-                clone: [
-                  {
-                    name: 'https',
-                    href: 'https://bitbucket.org/workspace/repo',
-                  },
-                ],
-              },
-            }),
-          );
-        },
-      ),
+    ).rejects.toThrow(
+      /Authorization has not been provided for no-credentials.bitbucket.com/,
     );
-
-    await action.handler({
-      ...mockContext,
-      input: {
-        ...mockContext.input,
-        repoUrl: 'bitbucket.org?workspace=workspace&project=project&repo=repo',
-      },
-    });
   });
 
-  it('should call the correct APIs when the host is hosted bitbucket', async () => {
+  it('should call the correct APIs with token', async () => {
     expect.assertions(2);
     server.use(
       rest.post(
         'https://hosted.bitbucket.com/rest/api/1.0/projects/project/repos',
         (req, res, ctx) => {
           expect(req.headers.get('Authorization')).toBe('Bearer thing');
-          expect(req.body).toEqual({ public: false, name: 'repo' });
+          expect(req.body).toEqual({
+            public: false,
+            name: 'repo',
+            defaultBranch: 'master',
+          });
           return res(
             ctx.status(201),
             ctx.set('Content-Type', 'application/json'),
@@ -203,14 +170,65 @@ describe('publish:bitbucket', () => {
     });
   });
 
-  it('should work if the token is provided through ctx.input', async () => {
+  it('should call the correct APIs with basic auth', async () => {
     expect.assertions(2);
     server.use(
       rest.post(
-        'https://notoken.bitbucket.com/rest/api/1.0/projects/project/repos',
+        'https://basic-auth.bitbucket.com/rest/api/1.0/projects/project/repos',
         (req, res, ctx) => {
-          expect(req.headers.get('Authorization')).toBe('Bearer lols');
-          expect(req.body).toEqual({ public: false, name: 'repo' });
+          expect(req.headers.get('Authorization')).toBe(
+            'Basic dGVzdC11c2VyOnRlc3QtcGFzc3dvcmQ=',
+          );
+          expect(req.body).toEqual({
+            public: false,
+            name: 'repo',
+            defaultBranch: 'master',
+          });
+          return res(
+            ctx.status(201),
+            ctx.set('Content-Type', 'application/json'),
+            ctx.json({
+              links: {
+                self: [
+                  {
+                    href: 'https://bitbucket.mycompany.com/projects/project/repos/repo',
+                  },
+                ],
+                clone: [
+                  {
+                    name: 'http',
+                    href: 'https://bitbucket.mycompany.com/scm/project/repo',
+                  },
+                ],
+              },
+            }),
+          );
+        },
+      ),
+    );
+
+    await action.handler({
+      ...mockContext,
+      input: {
+        ...mockContext.input,
+        repoUrl: 'basic-auth.bitbucket.com?project=project&repo=repo',
+      },
+    });
+  });
+
+  it('should work if the token is provided through ctx.input', async () => {
+    expect.assertions(2);
+    const token = 'user-token';
+    server.use(
+      rest.post(
+        'https://no-credentials.bitbucket.com/rest/api/1.0/projects/project/repos',
+        (req, res, ctx) => {
+          expect(req.headers.get('Authorization')).toBe(`Bearer ${token}`);
+          expect(req.body).toEqual({
+            public: false,
+            name: 'repo',
+            defaultBranch: 'master',
+          });
           return res(
             ctx.status(201),
             ctx.set('Content-Type', 'application/json'),
@@ -236,8 +254,9 @@ describe('publish:bitbucket', () => {
     await action.handler({
       ...mockContext,
       input: {
-        repoUrl: 'notoken.bitbucket.com?project=project&repo=repo',
-        token: 'lols',
+        ...mockContext.input,
+        repoUrl: 'no-credentials.bitbucket.com?project=project&repo=repo',
+        token: token,
       },
     });
   });
@@ -324,28 +343,37 @@ describe('publish:bitbucket', () => {
     });
   });
 
-  it('should call initAndPush with the correct values', async () => {
+  it('should call initAndPush with the correct values with token', async () => {
     server.use(
       rest.post(
-        'https://api.bitbucket.org/2.0/repositories/workspace/repo',
-        (_, res, ctx) =>
-          res(
-            ctx.status(200),
+        'https://hosted.bitbucket.com/rest/api/1.0/projects/project/repos',
+        (req, res, ctx) => {
+          expect(req.headers.get('Authorization')).toBe('Bearer thing');
+          expect(req.body).toEqual({
+            public: false,
+            name: 'repo',
+            defaultBranch: 'master',
+          });
+          return res(
+            ctx.status(201),
             ctx.set('Content-Type', 'application/json'),
             ctx.json({
               links: {
-                html: {
-                  href: 'https://bitbucket.org/workspace/repo',
-                },
+                self: [
+                  {
+                    href: 'https://bitbucket.mycompany.com/projects/project/repos/repo',
+                  },
+                ],
                 clone: [
                   {
-                    name: 'https',
-                    href: 'https://bitbucket.org/workspace/cloneurl',
+                    name: 'http',
+                    href: 'https://bitbucket.mycompany.com/scm/project/repo',
                   },
                 ],
               },
             }),
-          ),
+          );
+        },
       ),
     );
 
@@ -353,37 +381,107 @@ describe('publish:bitbucket', () => {
 
     expect(initRepoAndPush).toHaveBeenCalledWith({
       dir: mockContext.workspacePath,
-      remoteUrl: 'https://bitbucket.org/workspace/cloneurl',
+      remoteUrl: 'https://bitbucket.mycompany.com/scm/project/repo',
       defaultBranch: 'master',
-      auth: { username: 'x-token-auth', password: 'tokenlols' },
+      auth: { token: 'thing' },
       logger: mockContext.logger,
       commitMessage: 'initial commit',
-      gitAuthorInfo: {},
+      gitAuthorInfo: {
+        email: undefined,
+        name: undefined,
+      },
+    });
+  });
+
+  it('should call initAndPush with the correct values with basic auth', async () => {
+    server.use(
+      rest.post(
+        'https://basic-auth.bitbucket.com/rest/api/1.0/projects/project/repos',
+        (req, res, ctx) => {
+          expect(req.headers.get('Authorization')).toBe(
+            'Basic dGVzdC11c2VyOnRlc3QtcGFzc3dvcmQ=',
+          );
+          expect(req.body).toEqual({
+            public: false,
+            name: 'repo',
+            defaultBranch: 'master',
+          });
+          return res(
+            ctx.status(201),
+            ctx.set('Content-Type', 'application/json'),
+            ctx.json({
+              links: {
+                self: [
+                  {
+                    href: 'https://bitbucket.mycompany.com/projects/project/repos/repo',
+                  },
+                ],
+                clone: [
+                  {
+                    name: 'http',
+                    href: 'https://bitbucket.mycompany.com/scm/project/repo',
+                  },
+                ],
+              },
+            }),
+          );
+        },
+      ),
+    );
+
+    await action.handler({
+      ...mockContext,
+      input: {
+        ...mockContext.input,
+        repoUrl: 'basic-auth.bitbucket.com?project=project&repo=repo',
+      },
+    });
+
+    expect(initRepoAndPush).toHaveBeenCalledWith({
+      dir: mockContext.workspacePath,
+      remoteUrl: 'https://bitbucket.mycompany.com/scm/project/repo',
+      defaultBranch: 'master',
+      auth: { username: 'test-user', password: 'test-password' },
+      logger: mockContext.logger,
+      commitMessage: 'initial commit',
+      gitAuthorInfo: {
+        email: undefined,
+        name: undefined,
+      },
     });
   });
 
   it('should call initAndPush with the correct default branch', async () => {
     server.use(
       rest.post(
-        'https://api.bitbucket.org/2.0/repositories/workspace/repo',
-        (_, res, ctx) =>
-          res(
-            ctx.status(200),
+        'https://hosted.bitbucket.com/rest/api/1.0/projects/project/repos',
+        (req, res, ctx) => {
+          expect(req.headers.get('Authorization')).toBe('Bearer thing');
+          expect(req.body).toEqual({
+            public: false,
+            name: 'repo',
+            defaultBranch: 'main',
+          });
+          return res(
+            ctx.status(201),
             ctx.set('Content-Type', 'application/json'),
             ctx.json({
               links: {
-                html: {
-                  href: 'https://bitbucket.org/workspace/repo',
-                },
+                self: [
+                  {
+                    href: 'https://bitbucket.mycompany.com/projects/project/repos/repo',
+                  },
+                ],
                 clone: [
                   {
-                    name: 'https',
-                    href: 'https://bitbucket.org/workspace/cloneurl',
+                    name: 'http',
+                    href: 'https://bitbucket.mycompany.com/scm/project/repo',
                   },
                 ],
               },
             }),
-          ),
+          );
+        },
       ),
     );
 
@@ -397,30 +495,29 @@ describe('publish:bitbucket', () => {
 
     expect(initRepoAndPush).toHaveBeenCalledWith({
       dir: mockContext.workspacePath,
-      remoteUrl: 'https://bitbucket.org/workspace/cloneurl',
+      remoteUrl: 'https://bitbucket.mycompany.com/scm/project/repo',
       defaultBranch: 'main',
-      auth: { username: 'x-token-auth', password: 'tokenlols' },
+      auth: { token: 'thing' },
       logger: mockContext.logger,
       commitMessage: 'initial commit',
-      gitAuthorInfo: {},
+      gitAuthorInfo: {
+        email: undefined,
+        name: undefined,
+      },
     });
   });
 
   it('should call initAndPush with the configured defaultAuthor', async () => {
     const customAuthorConfig = new ConfigReader({
       integrations: {
-        bitbucket: [
-          {
-            host: 'bitbucket.org',
-            token: 'tokenlols',
-          },
+        bitbucketServer: [
           {
             host: 'hosted.bitbucket.com',
             token: 'thing',
             apiBaseUrl: 'https://hosted.bitbucket.com/rest/api/1.0',
           },
           {
-            host: 'notoken.bitbucket.com',
+            host: 'no-credentials.bitbucket.com',
           },
         ],
       },
@@ -434,32 +531,41 @@ describe('publish:bitbucket', () => {
 
     const customAuthorIntegrations =
       ScmIntegrations.fromConfig(customAuthorConfig);
-    const customAuthorAction = createPublishBitbucketAction({
+    const customAuthorAction = createPublishBitbucketServerAction({
       integrations: customAuthorIntegrations,
       config: customAuthorConfig,
     });
 
     server.use(
       rest.post(
-        'https://api.bitbucket.org/2.0/repositories/workspace/repo',
-        (_, res, ctx) =>
-          res(
-            ctx.status(200),
+        'https://hosted.bitbucket.com/rest/api/1.0/projects/project/repos',
+        (req, res, ctx) => {
+          expect(req.headers.get('Authorization')).toBe('Bearer thing');
+          expect(req.body).toEqual({
+            public: false,
+            name: 'repo',
+            defaultBranch: 'master',
+          });
+          return res(
+            ctx.status(201),
             ctx.set('Content-Type', 'application/json'),
             ctx.json({
               links: {
-                html: {
-                  href: 'https://bitbucket.org/workspace/repo',
-                },
+                self: [
+                  {
+                    href: 'https://bitbucket.mycompany.com/projects/project/repos/repo',
+                  },
+                ],
                 clone: [
                   {
-                    name: 'https',
-                    href: 'https://bitbucket.org/workspace/cloneurl',
+                    name: 'http',
+                    href: 'https://bitbucket.mycompany.com/scm/project/repo',
                   },
                 ],
               },
             }),
-          ),
+          );
+        },
       ),
     );
 
@@ -467,8 +573,8 @@ describe('publish:bitbucket', () => {
 
     expect(initRepoAndPush).toHaveBeenCalledWith({
       dir: mockContext.workspacePath,
-      remoteUrl: 'https://bitbucket.org/workspace/cloneurl',
-      auth: { username: 'x-token-auth', password: 'tokenlols' },
+      remoteUrl: 'https://bitbucket.mycompany.com/scm/project/repo',
+      auth: { token: 'thing' },
       logger: mockContext.logger,
       defaultBranch: 'master',
       commitMessage: 'initial commit',
@@ -479,54 +585,59 @@ describe('publish:bitbucket', () => {
   it('should call initAndPush with the configured defaultCommitMessage', async () => {
     const customAuthorConfig = new ConfigReader({
       integrations: {
-        bitbucket: [
-          {
-            host: 'bitbucket.org',
-            token: 'tokenlols',
-          },
+        bitbucketServer: [
           {
             host: 'hosted.bitbucket.com',
             token: 'thing',
             apiBaseUrl: 'https://hosted.bitbucket.com/rest/api/1.0',
           },
           {
-            host: 'notoken.bitbucket.com',
+            host: 'no-credentials.bitbucket.com',
           },
         ],
       },
       scaffolder: {
-        defaultCommitMessage: 'Test commit message',
+        defaultCommitMessage: 'initial commit',
       },
     });
 
     const customAuthorIntegrations =
       ScmIntegrations.fromConfig(customAuthorConfig);
-    const customAuthorAction = createPublishBitbucketAction({
+    const customAuthorAction = createPublishBitbucketServerAction({
       integrations: customAuthorIntegrations,
       config: customAuthorConfig,
     });
 
     server.use(
       rest.post(
-        'https://api.bitbucket.org/2.0/repositories/workspace/repo',
-        (_, res, ctx) =>
-          res(
-            ctx.status(200),
+        'https://hosted.bitbucket.com/rest/api/1.0/projects/project/repos',
+        (req, res, ctx) => {
+          expect(req.headers.get('Authorization')).toBe('Bearer thing');
+          expect(req.body).toEqual({
+            public: false,
+            name: 'repo',
+            defaultBranch: 'master',
+          });
+          return res(
+            ctx.status(201),
             ctx.set('Content-Type', 'application/json'),
             ctx.json({
               links: {
-                html: {
-                  href: 'https://bitbucket.org/workspace/repo',
-                },
+                self: [
+                  {
+                    href: 'https://bitbucket.mycompany.com/projects/project/repos/repo',
+                  },
+                ],
                 clone: [
                   {
-                    name: 'https',
-                    href: 'https://bitbucket.org/workspace/cloneurl',
+                    name: 'http',
+                    href: 'https://bitbucket.mycompany.com/scm/project/repo',
                   },
                 ],
               },
             }),
-          ),
+          );
+        },
       ),
     );
 
@@ -534,8 +645,8 @@ describe('publish:bitbucket', () => {
 
     expect(initRepoAndPush).toHaveBeenCalledWith({
       dir: mockContext.workspacePath,
-      remoteUrl: 'https://bitbucket.org/workspace/cloneurl',
-      auth: { username: 'x-token-auth', password: 'tokenlols' },
+      remoteUrl: 'https://bitbucket.mycompany.com/scm/project/repo',
+      auth: { token: 'thing' },
       logger: mockContext.logger,
       defaultBranch: 'master',
       commitMessage: 'initial commit',
@@ -546,25 +657,34 @@ describe('publish:bitbucket', () => {
   it('should call outputs with the correct urls', async () => {
     server.use(
       rest.post(
-        'https://api.bitbucket.org/2.0/repositories/workspace/repo',
-        (_, res, ctx) =>
-          res(
-            ctx.status(200),
+        'https://hosted.bitbucket.com/rest/api/1.0/projects/project/repos',
+        (req, res, ctx) => {
+          expect(req.headers.get('Authorization')).toBe('Bearer thing');
+          expect(req.body).toEqual({
+            public: false,
+            name: 'repo',
+            defaultBranch: 'master',
+          });
+          return res(
+            ctx.status(201),
             ctx.set('Content-Type', 'application/json'),
             ctx.json({
               links: {
-                html: {
-                  href: 'https://bitbucket.org/workspace/repo',
-                },
+                self: [
+                  {
+                    href: 'https://bitbucket.mycompany.com/projects/project/repos/repo',
+                  },
+                ],
                 clone: [
                   {
-                    name: 'https',
-                    href: 'https://bitbucket.org/workspace/cloneurl',
+                    name: 'http',
+                    href: 'https://bitbucket.mycompany.com/scm/project/repo',
                   },
                 ],
               },
             }),
-          ),
+          );
+        },
       ),
     );
 
@@ -572,36 +692,45 @@ describe('publish:bitbucket', () => {
 
     expect(mockContext.output).toHaveBeenCalledWith(
       'remoteUrl',
-      'https://bitbucket.org/workspace/cloneurl',
+      'https://bitbucket.mycompany.com/scm/project/repo',
     );
     expect(mockContext.output).toHaveBeenCalledWith(
       'repoContentsUrl',
-      'https://bitbucket.org/workspace/repo/src/master',
+      'https://bitbucket.mycompany.com/projects/project/repos/repo',
     );
   });
 
   it('should call outputs with the correct urls with correct default branch', async () => {
     server.use(
       rest.post(
-        'https://api.bitbucket.org/2.0/repositories/workspace/repo',
-        (_, res, ctx) =>
-          res(
-            ctx.status(200),
+        'https://hosted.bitbucket.com/rest/api/1.0/projects/project/repos',
+        (req, res, ctx) => {
+          expect(req.headers.get('Authorization')).toBe('Bearer thing');
+          expect(req.body).toEqual({
+            public: false,
+            name: 'repo',
+            defaultBranch: 'main',
+          });
+          return res(
+            ctx.status(201),
             ctx.set('Content-Type', 'application/json'),
             ctx.json({
               links: {
-                html: {
-                  href: 'https://bitbucket.org/workspace/repo',
-                },
+                self: [
+                  {
+                    href: 'https://bitbucket.mycompany.com/projects/project/repos/repo',
+                  },
+                ],
                 clone: [
                   {
-                    name: 'https',
-                    href: 'https://bitbucket.org/workspace/cloneurl',
+                    name: 'http',
+                    href: 'https://bitbucket.mycompany.com/scm/project/repo',
                   },
                 ],
               },
             }),
-          ),
+          );
+        },
       ),
     );
 
@@ -615,11 +744,11 @@ describe('publish:bitbucket', () => {
 
     expect(mockContext.output).toHaveBeenCalledWith(
       'remoteUrl',
-      'https://bitbucket.org/workspace/cloneurl',
+      'https://bitbucket.mycompany.com/scm/project/repo',
     );
     expect(mockContext.output).toHaveBeenCalledWith(
       'repoContentsUrl',
-      'https://bitbucket.org/workspace/repo/src/main',
+      'https://bitbucket.mycompany.com/projects/project/repos/repo',
     );
   });
 });
