+++ plugins/scaffolder-backend-module-bitbucket/src/actions/bitbucketServerPullRequest.test.ts
@@ -14,8 +14,9 @@
  * limitations under the License.
  */
 
-jest.mock('../helpers', () => {
+jest.mock('@backstage/plugin-scaffolder-node', () => {
   return {
+    ...jest.requireActual('@backstage/plugin-scaffolder-node'),
     initRepoAndPush: jest.fn().mockResolvedValue({
       commitHash: '220f19cc36b551763d157f1b5e4a4b446165dbd6',
     }),
